Revert D54856044: Multisect successfully blamed "D54856044: [Thrift] Unify op and type library targets" for one build failure
Summary:
This diff reverts D54856044
D54856044: [Thrift] Unify op and type library targets by vitaut causes the following build failure:

Tests affected:
- [aloha_nativelibs_rdk_x86](https://www.internalfb.com/intern/test/844424995625643/)

Here's the Multisect link:
https://www.internalfb.com/multisect/4597066
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: iahs

Differential Revision: D54903562

fbshipit-source-id: e20ec9b3c75f3b3d61a9ec9bb0fbc4386cf9e4c8