Enable support for MuxIOThreadPoolExecutor::setNumThreads()
Summary:
`MuxIOThreadPoolExecutor:setNumThreads()` works as expected as long as we don't pass 0 as argument: that would cause the executor to deadlock because no thread is left driving the eventbases.

This diff disallows that case (and also the case of #threads > #evbs, as it is likely indicative of misconfiguration), and adds tests verifying the basic functionality.

Reviewed By: dmm-fb

Differential Revision: D54857899

fbshipit-source-id: 0bd2d5441a13f0fa847f4dece4f16a6405efd5fc