Partition Abstraction
Summary:
# Context
The subtype relation for `TypeSwitch` which is introduced in D54664466 will require partition types into two disjoint types. To support this logic I pulled out the partitioning logic into its own module that can be tested.

A partition is represented as a triple of (left, span, right). The left is everything that is on the left side of the partition, right is everything that is on the right side of the partition, while span represents entries that are partially over both the left and right sides of the partition.

Partitions can be joined or intersected that will itself produce another partition. When they do their subsequent left, span and right components will be joined or intersected.

# Code decisions
I choose to keep the unions and intersections in disjunctive normal form. I use a GADT to validate we keep to this structure.

Reviewed By: enetsee

Differential Revision: D54777531

fbshipit-source-id: 0d6593bcb33649d8521bb4b553291ff6c560eef9