don't abuse Pos.none in override_method quickfix
Summary:
before: create a quickfix edit with nonsensical position data. From the user's point of view (if this code is reachable at all) we probably either crashed or produced a weird edit near the beginning of the file.

after: don't create a quickfix in such situations, instead log an invariant violation bug and use an empty edit list

Differential Revision: D54949365

fbshipit-source-id: 2b2b2f5c575200d863e33dfe3b65ef36a55f4c3e