Change surface of meth_caller intrinsic
Summary:
Internalize the `string` representation of the meth_caller class to systemlib, letting class pointers and lazy classes flow in without implicit conversion.

The overall goal is for `meth_caller` to carry a class pointer while `dyn_meth_caller` carries a string. However, updating the `$class` property on `MethCallerHelper` is a rabbit hole due to JIT functions that assert the type of the prop and templatized shared implementations for dyn_meth_caller+meth_caller; moving that effort to {D54882921}.

Reviewed By: jano

Differential Revision: D54882879

fbshipit-source-id: e59c1466034453f38c4858c8eaa9ad7988bcc29a