split Param/DefaultValue
Summary:
Each param can have an optional DefaultValue, and the DefaultValue specifies
the function entry point to use to initialize missing default values at runtime.
ir and hhbc params are the same except for how they specify the entry point.

To unblock unification, factor DefaultValue out of Param, and express params
as Vec<(Param, Option<DefaultValue>)>.  In hhbc we use a new ParamEntry struct
with named fields to avoid tuples in the FFI.

Reviewed By: aorenste

Differential Revision: D54484147

fbshipit-source-id: 74640ce783d40ad0d12bab5ec30661b3247fd04e