Use ArrayKeyTypes by ArrayIterProfile
Summary:
When profiling iterators, we first reduce the 4 bit ArrayKeyTypes information
to 2 bits, losing potentially valuable information, such as Int|StaticStr.

Then we count how many times did we encounter each of these 4 possible values.

Finally, if only one of these values were seen, we pick that type, otherwise we
pick the most generic ArrayKey.

None of this complexity is required. Replace it with just keeping track of
the observed ArrayKeyTypes bits.

Improves the key type inference and brings us a step closer to be able to
handle tombstones in specialized loops.

Reviewed By: ricklavoie

Differential Revision: D54977210

fbshipit-source-id: aa560b49e8c236db5fe1e410d79a4d0ee85e5657