AllocationColocator: allow void template param to mean no root object
Summary:
The added doc block explains it well. After this diff `void` has a special meaning — it's effectively considered a zero-sized type.

As such, the callback passed to `allocate` does not require the user to return `T` if `T` is void. The rest of the changes are mostly dealing with the fact that `void` is incomplete — `sizeof`, `alignof`, invoking destructor etc.

Meta:
This will let us use `AllocationColocator` in `RequestsRegistry::DebugStub` where the ownership of objects is highly complex.

https://www.internalfb.com/code/fbsource/[4fff99c46e24]/fbcode/thrift/lib/cpp2/server/RequestsRegistry.h?lines=262-270
https://www.internalfb.com/code/fbsource/[4fff99c46e24]/fbcode/thrift/lib/cpp2/server/RequestsRegistry.cpp?lines=261-266

Reviewed By: sethdelliott

Differential Revision: D54883424

fbshipit-source-id: d31a8ba05892debfe6f2011daf3063a5ef97018e