Make RequestsRegistry::DebugStub use AllocationColocator
Summary:
This diff just removes the ugly pointer arithmetic in `DebugStub` with `AllocationColocator`. The behavior and layout in memory remains the same.

The change in `InstrumentationTest.cpp` is putting `ServerConfigsMock` on the heap because (for some reason) `alignof(ServerConfigsMock)` is 64 on my machine. That is of course higher than the maximum allowed by `AllocationColocator`:
https://www.internalfb.com/code/fbsource/[0ca24769888f]/fbcode/thrift/lib/cpp2/util/AllocationColocator.h?lines=451-461

Thankfully, none of the `ResponseChannelRequest` objects in our actual production code have this problem. Notice that everything compiles. Here are the relevant subclasses
```
apache::thrift::ThriftRequestCore
  apache::thrift::ThriftRequest
  apache::thrift::rocket::RocketThriftRequest
    apache::thrift::rocket::ThriftServerRequestResponse
    apache::thrift::rocket::ThriftServerRequestFnf
    apache::thrift::rocket::ThriftServerRequestStream
    apache::thrift::rocket::ThriftServerRequestSink
```

Reviewed By: sethdelliott

Differential Revision: D54921972

fbshipit-source-id: 3508f19eac397b648b479f9553fe6d1dca09bfb0