Allow collection syntax in expression trees
Summary:
Allow C { k => v, ...} syntax inside of expression trees. Here C is not restricted to `Map` as in Hack generally. The runtime semantics is to generate a call to `visitCollection`
which is a function `function(?ExprPos $_, string $collection, (t, t) ...$operand): t)` that will be passed the name `C` of the collection, and all of the visited key/value pairs. The client type is determined by a call to `C` which should have type `function<Tk as ClientKeyType, Tv as ClientValueType>((Tk, Tv) ...$_)): ClientMapType<Tk, Tv>. Both `visitCollection` and `C` should be functions in the expression tree's class.

We don't specify what the collection can be called to allow expression tree clients to choose domain specific terminology.

This is an unstable feature for now

Reviewed By: dlreeves

Differential Revision: D51987405

fbshipit-source-id: 3c094f5e70d6fae28216868e0fb0e8e21e2c6660