Don't call back into `default_subtype` with `ConstraintType`s - for `Can_traverse`
Summary:
In order to remove internal_ty we need to avoid lifting non-subtype constraints back into `Subtype`, via `default_subtype` or `simplify_subtype_i`, for the common var/union/intersection/generic of lhs cases.

This diff does so for `Can_traverse` constraints. The one interesting thing to note is that, since it used to recurse through `simplify_subtype_i` we have to generate a new `Typing_error.t` at every recursive call to maintain existing behavior. This is almost certainly a bug but is better dealt with in a subsequent diff

Differential Revision: D54804901

fbshipit-source-id: 568b7659a0737a50e100ae457f719c61fc50c8b2