Implement Subtyping for Ttype_switch
Summary:
# Context
This diff stack implements the new constraint based type refinement end to end when typing the statement:
```
if (_ is bool) {
  ...
} else {
  ...
}
```
We ultimately will utilize the constraint type added in D54664466, which provides more context for this stack.

# This Diff
We implement the subtyping rule for `Ttype_switch`. The declarative typing rule is:

```
T <: TypeSwitch(P, T1, T2)
T1 <: A      T2 <: B
----------------------------
T <: TypeSwitch(P, A, B)
```
In order to make this declarative rule, algorithmic we utilize the type partitioning developed in D54866425. By computing `ty_partition(P, T)` we get back a triple of types `(T_left, T_span, T_right)`. We then have:
```
T1 = T_left | (T_span & Top_P)
T2 = T_right | (T_span & Top_!P)
```

## Implementation Details
`T_left`, `T_span` and `T_right` respectively are in disjunctive normal form, encoded as the type `locl_ty list list`. We avoid calling into `Typing_intersection` or `Typing_union` since simplifying the types can lead to divergent behavior in some instances.

For unions this is easy to avoid since we end up producing conjunctive propositions, i.e. if `T_left = U | V` then `T_left <: A` simplifies to `U <: A &&& V <: A`.

For intersections we directly construct `Tintersection` instances. The intersections will not be simplified, but subtyping should not depend on simplification to be correct. In practice though there is a risk this can lead to incompleteness in our current implementation. I believe the risk of this is low given the end to end implementation did not reveal any new errors detected across our codebase.

The last thing to detail is how `Top_P` and `Top_!P` are encoded. For `Top_!P` this is simple. We construct a `Tneg (Neg_predicate P)` which should work for any predicate. For `Top_P` though its representation will be dependent on `P`. We can deal with this as we move more of refinement to use this system, but for the MVP I manually do the translation of `IsBool` to the type `Tprim Tbool`.

# This Stack
1. Implement Type Partitioning
2. **Implement Subtyping for Ttype_switch**
3. Use Ttype_switch to handle refinements for the statement `if(_ is bool)`

Differential Revision: D54866426

fbshipit-source-id: 3ac3baf20a6b9d7dace72b5df3e483a888185fd5