Switch if(_ is bool) to use Constraint-Based Refinement
Summary:
# Context
This diff stack implements the new constraint based type refinement end to end when typing the statement:
```
if (_ is bool) {
  ...
} else {
  ...
}
```
We ultimately will utilize the constraint type added in D54664466, which provides more context for this stack.

# This Diff
After subtyping is implemented for `Ttype_switch` we can now utilize the new constraint type. For this we update `typing.ml` , when typing the statement `If(expr, branch1, branch2)` we check `expr` to see if the `if` statement can be treated like a type switch using `Typing.Eval.type_switch`

`Typing.Eval.type_switch` currently looks for expressions of the form `expr is Hint` and will produce a type switch if both `expr` can be interpreted as a local variable and if the `Hint` can be interpreted as a type predicate. If so we open a new scope for type variables and introduce two fresh type variables to represent the type of `expr` when the predicate is true and the other when it is false. We then add the constraint

```
  expr: T <: TypeSwitch(Hint, ty_true, ty_false)
```
then close and solve the scope we opened for the type variables.

## Implementation Details

There was a small change made to `Typing_union` to allow `bool | not bool` to be simplified. This was handled in a generic way where `T | Tneg (Neg_predicate P)` is simplified to `mixed` if we can transform `T` to another predicate `P_T` and `P = P_T`.

# This Stack
1. Implement Type Partitioning
2. Implement Subtyping for Ttype_switch
3. **Use Ttype_switch to handle refinements for the statement `if(_ is bool)`**

Differential Revision: D54866428

fbshipit-source-id: f4f190b202fcc35148e9a282de30896803dbdc59