Implement Type Partitioning
Summary:
# Context
This diff stack implements the new constraint based type refinement end to end when typing the statement:
```
if (_ is bool) {
  ...
} else {
  ...
}
```
We ultimately will utilize the constraint type added in D54664466, which provides more context for this stack.

# This Diff
Before defining subtyping for the `Ttype_switch` constraint type we need to introduce a new operation on types that I am dubbing **type partitioning**. Type partitioning involves splitting a type into three types based on its relation to a given type predicate. For a given predicate `P`, a type `T` is either:
* **Left** of the predicate if `T <: Top_P`
* **Right** of the predicate if `T <: Top_!P`
* Otherwise `T` is **Spans** the predicate, meaning some values in `T` are in `Top_P` and some are in `Top_!P`

Partitions are algebraic, supporting unions and intersections to build up other partitions. This algebra is implemented in D54777531. Using this algebra we can implement a function, `split(P, T)`, that given a type and type predicate outputs a partition of types over the given predicate.

This splitting function generally has four cases:
1. If a type `T` can be decomposed into a union of types `T1 | T2`, then `split(P, T) = split(P, T1) ∪ split(P, T2)`
2. If a type `T` can be decomposed into an intersection of type `T1 & T2`, then `split(P, T) = split(P, T1) ∩ split(P, T2)`
3. If a type `T` is an abstract type with an upper bound of `U`, then `split(P, T) = span(T) ∩ split(P, U)`
4. Otherwise `T` cannot be decomposed further

When the last case is hit we can determine if `T` is Left, Right, or Span by translating `T`, `Top_P` and `Top_!P` to data types can see if they are disjoint.

## Implementation Details
The above algorithm is implemented in a new file `Typing_refinement`. It utilizes `Typing_case_types` to reason about the relationship of types to type predicates in terms of their runtime data type. This requires exposing more internals of `Typing_case_types`. To minimize the amount of information leaked, I add a new module `Typing_case_types.AtomicDataType` that translates data types for a subset of possible `locl_ty`s that are considered atomic and cannot be decomposed further.

Finally I made a small extension to `Partition` for simplifying the DNF formulas produced. For instance consider the type `T as bool` and computing `split(IsBool, T)`. Based on the algorithm we would produce a partition `Left(T & int)`. This is correct but the intersection can be simplified to only `T`. The same applies for a type that falls fully under Right or Span.

# This Stack
1. **Implement Type Partitioning**
2. Implement Subtyping for Ttype_switch
3. Use Ttype_switch to handle refinements for the statement `if(_ is bool)`

Differential Revision: D54866425

fbshipit-source-id: d89de50fb56a5bf5ffabbd4feb3eca98f6938e7f