Add `hint_style` to `Quickfix.t`
Summary:
It is my intention to make certain quickfixes available when the user has their cursor on positions that are __not__ the primary error position.

For this, I add the type `Quickfix.hint_style` and add a field `hint_styles` to `Quickfix.t`, where we can specificy additional locations
where the quickfix should be made available.

## This diff

- [x] Introduces `Classish_positions_types`, factoring out type definitions previously in `Classish_positions` into a separate `.ml` file. This is **to work around oxidized limitations.**
- [x] In `quickfix.ml`: add the `hint_styles` field to `Quickfix.t`
- [x] Rename `Quickfix.make_eager` to `Quickfix.make_eager_primary` to indicate that it will only show hints where the primary error is located
  - **Note:** in a subsequent diff, I change this again...
- [x] Callsites are updated, but the diff is functionally a noop

Reviewed By: mheiber

Differential Revision: D54984464

fbshipit-source-id: 6c6111fdd856062585f1dcf39de9511e23ea69a5