fix hhi for `module_exists` to match reality
Summary:
D39736461 implemented this as a global namespace function, but D52627600 added it in the HH namespace. We could fix by either moving the runtime to the HH namespace (and wait for an hhvm push) or by fixing up the hhi files to match reality (and waiting for a hack push).

Given that `HH\package_exists` is already working e2e in the HH namespace, we've here opted for matching `HH\module_exists` (ie: fixing the runtime to match the hhi) for consistency. This means we're waiting for an hhvm push intead of just a hack push to roll this out.

Reviewed By: DavidSnider

Differential Revision: D54971655

fbshipit-source-id: db19284be9a948d88e54bf8f45bafe2e2f580e90