SBE response encoding prototype
Summary:
Adding a new prototype API for bypass RPC to test the new SBE encoding that the thrift team is working on.

For SBE encoding, the data types and message formats are declared in the new MySQLTypes.xml (this is the equivalent of the thrift data structures declared in .thrift files). All the files under the `mysql_thrift_server/if/apache_thrift_sbe` are generated from the XML by running `buck2 run thrift/vendor/simple-binary-encoding/sbe-tool:sbe_tool -- -J-Dsbe.target.language=Cpp -J-Dsbe.generate.precedence.checks=true -J-Dsbe.output.dir=mysql_thrift_server/if mysql_thrift_server/if/MySQLTypes.xml`.

NOTE: The MySQLTypes.xml is just added as part of the prototype, but eventually, it would be generated during builds out of the .thrift files.

In this diff, I'm just encoding the response struct into SBE, and added a test that round-trips a `MySQLResponse` thrift struct into SBE and back to thrift to validate the encoding correctness.

The response contains the following components:
- `MySQLResponseHeader` message: which contains the list of column names and data types
- For each column in each row, one of MySQLNull/MySQLBool/MySQLUInt/... messages
- A `MySQLResponseEnd` marking the end of row data.

Reviewed By: jkedgar

Differential Revision: D53869829

fbshipit-source-id: b16ebbc1a499c7ee4509fb12686c0babe396705e