Add async_create{Interaction} method wrapper for IG thrift
Summary:
IG has wrapper thrift clients (Py3ThriftClientCounterWrapper) that only takes async functions. In thrift-py3, interactions are sync, and don't work with the wrapper. The proper thing to do is migrate to thrift-python, unfortunately, that is very difficult at the moment.

This diff adds a wrapper around the sync method so that it can be used in ig.

Reviewed By: nanshu

Differential Revision: D55030773

fbshipit-source-id: 0220be6f3c3c0b6aeebe2f7f5c15faafbf5ee3d1