Remove redundant struct Adata
Summary:
HHBC instructions refer to vec, dict, and keyset literals with an integer AdataId.
The literals are stored in a Vec<Adata> where Adata is an {AdataId, TypedValue} pair.
The Adata::id struct is redundant because AdataIds are integer numbers 0..N
that can be used as an index. Since the only other field is the TypedValue literal,
the struct itself is redundant.

The actual AdataId numbers have no semantic meaning other than for locating
the associated literal, so it's not important to preserve them in round trips,
keep dead values around, etc.

Drop the redundant struct & id field.
Modify bytecode printer/assembler to assign numbers on output and
remap numbers on input.

hackc-translator already remapped ids when inserting literals into FuncEmitter.

Reviewed By: aorenste

Differential Revision: D54984638

fbshipit-source-id: afc69dd75a3428bc3c672feb0d57adf4cec19962