Improve refinement
Summary:
When checking a lambda that is immediately called, don't invalidate the fake members, when the arguments are syntactically constant. This is sound because there's nothing that can happen between making the lambda and calling it that might invalidate them. This supports the pattern where an immediately invoked lambda is used to inject statements into an expression, as in `async` blocks and the runtime expression of an expression tree.

Ideally, we would defer typing the lambda until after checking the non-lambda arguments, and then we wouldn't have to be conservative about the arguments being syntactic constants. However, this would be a big refactor to how the code works now.

Reviewed By: madgen

Differential Revision: D54951550

fbshipit-source-id: a41a5341e98b0809c889b77eb36b98418ee293b1