Don't call into `simplfy_subtype_i` during transitivity step
Summary:
The aim of this stack is to eliminate the use of `internal_type` in constraint simplification. `internal_type` was used to represent both subtype constraints and all our other top-level constraints and dispatched to the appropriate handler from `simplify_subtype_i`.

This diff replaces calls to `simplify_subtype_i` during our 'transitivity' step (i.e. when we add new bounds to a tyvar and discover new constraints) with corresponding calls to each top-level constraint handler.

The change makes explicit our invariant that non-subtype constraints can only be 'upper bounds' (though not yet impossible) and that the consequence of adding a subtype constraint upper-bound can only ever be a change in the set of subtype upper-bounds (i.e. it can't result in a new non-subtype constraint)

The curious thing about this diff is in the case of adding a new non-subtype constraint. Because of the way we attempt to eliminate redundant bounds  (with `try_intersect_i`) adding a non-subtype constraint can result in modified _subtype_ constraints. This is likely a bug in the upper-bounds simplification mechanism.

The diff has a positive line count since we can't remove `internal_type` handling completely just yet.

Reviewed By: andrewjkennedy

Differential Revision: D55260052

fbshipit-source-id: 90059f1900e6e8d682c4390323debea880a0ee83