Fix CgroupResource's CPU percentile issue
Summary:
We have percentile stats implemented in `HostResources` but those not defined in `CgroupResources`

In this diff I'm trying to fill that GAP otherwise things won't work for targeting.

We need to have two thing:
- percentile target
- actual CPU percentile metric

As a percentile target I'm using callback provided to `HostResources` and as a percentile metric I'm using CPU utilisation.

Reviewed By: meleshuk

Differential Revision: D54676137

fbshipit-source-id: 4179c1f147f9d919f0dc1eb766600b8abd2cfb5b