Don't look for closure vars in typechecker-only code
Summary: hackc uses a visitor to compute the free variables of a lambda to know which ones go into the closure. That visitor was entring into the type-checker-only fields of ExpressionTree expressions. This made it inappropriately put temporary variables for the splices (`$0splice0`) in the closure. This updates the computation to only consider the `runtime` part of the ExpressionTree expression.

Reviewed By: viratyosin

Differential Revision: D55434735

fbshipit-source-id: ce78acd8a56f06e6b3566a50edbed9b0f89d73d1