Remove cached ItemView from FrozenRange iterator
Summary:
This uses an "arrow proxy" to avoid needing to eagerly update
an ItemView in the FrozenRange iterator, making the type smaller and
saving work so long as the workload doesn't dereference the same
iterator multiple times in such a way that the compiler can't figure
out that the two dereferences return the same result.

RFC status because this broke builds in FrozenBench, so I assume it
breaks builds elsewhere too and we'll want to be sure we want this
before spending time unbreaking builds. We like performance though,
right?

Reviewed By: ddrcoder

Differential Revision: D54706879

fbshipit-source-id: ec591e249796456adc06ad39a69db6e4ecd8f0b0