Remove the need to ReplyState to be wrapped in a Mutex
Summary:
We wrap the underlying C++ ReplyState in a mutex instead, so the parent class having a mutex is useless.

This lets us pass around the C++ reply state for sending replies back without having to clone the executor which saves a lot of overhead.

This diff also adjusts the queue timeout cogwheel test + maps a incorrectly set default value for the queueing timeout.

Reviewed By: Imxset21

Differential Revision: D55493894

fbshipit-source-id: 6a1c2da50903eab2ab5c6e100610212e00081a4c