Allow function decls with different case
Summary:
Since functions are case sensitive now, this should be no problem.

Also, file boundaries generally aren't supposed to be meaningful in hack, except
for the scope of file attributes, newtype, and the lower bound on module granularity.

If we continue to ban global (or package/deployment scoped) case colliding function
names, it should be done using providers that have access to global names.

Reviewed By: vassilmladenov

Differential Revision: D54581058

fbshipit-source-id: 17d8215835ab2350d175bd01ca5c83d443a080a9