use DeclMap and hhbc::Local
Summary:
One place, we had a `StringIdMap<u32> that could/should have been a DeclMap.

And DeclMap should be a `StringIdMap<hhbc::Local>`.  we have newtype wrappers for a good reason.

While I was here, throw an error if there are repeated local or parameter names. hackc is supposed to handle renaming them before they ever reach HHAS so this seems reasonable.

Reviewed By: aorenste

Differential Revision: D55128258

fbshipit-source-id: be91f1fd9d92b14976dc76336cf94c755b920f67