encapsulate newtypes
Summary:
It bothered me that LocId and IterId newtypes had a public `idx` field so I made it private and tidied up the accessors.
Callers typically need to use these as indices, which can be
done safely with IdVec, or using helpers that give you `usize`.

the internal `u32` is better encapsulated now.

Reviewed By: aorenste

Differential Revision: D55194045

fbshipit-source-id: 1f2fc8954f5a94e9bcd6ff19fb0f6d98ce067950