Add nocase support, bulk insert, and TypedDb<K> abstraction
Summary:
Some upgrades to Rust LMDB bindings I need later in the stack
* Introduce Snapshot, a version of RoTxn meant to be longer lived
and safe to pass between threads or store in a threadlocal. RoTxn can only
live as long as the Env it belongs to, enforced by rust borrow checker,
and makes sense for short lived transactions (the txn nests inside the
callstack that owns an Env).
Snapshot keeps it's Env alive via refcounting; it is useful for ownerhip
patterns that can't statically check RoTxn<->Env lifetime constraints,
such as having one Snapshot per HHVM request.

* Support for `nocase` keys, required for case-insensitve type lookups
when used for an HHVM FactsDB prototype later in in the stack.

* Additional bindings for bulk_append() and key-prefix iteration.

* bugfix: Rust iterator should own a MDB_cursor because stepping
mutates the cursor. the actual iterator state is the cursor.
* a new unit test

And a higher level TypedDb<K> abstraction inspired by `heed`.
TypedDb intends to be a zero-cost abstraction to allow zero copy type
safety for keys (and later values). A client can define a key type with primitives
to view or create keys as &[u8] or &K, and encapsulate primitive operations
such as key comparison in one place.

Later we could/should do the same thing for values, so clients can
have a TypedDb<K,V>, similar to a BTreeMap<K,V>.

Reviewed By: chadaustin

Differential Revision: D51719204

fbshipit-source-id: 164cfbf1d0b2f6efc6bdc4ea2bda0aa4ce470612