Function specializations for big-but-trivial callables
Summary:
When the callable is big but trivially-copyable, we can share destroy operations between all instances. There is not even any need to split them by size, as is done for small trivially-copyable callables, since there is extra space in the buffer to store the size and even the alignment.

As hypothesis, many big callables stored in a `Function` are also trivially-copyable, storing mostly references but also and pointers and arithmetic types.

Reviewed By: Gownta

Differential Revision: D32977854

fbshipit-source-id: 270a0dbd8ad2d7105ed567d3cf228045d7e7270f