document exception difference between py3 and python
Summary:
These tests document a difference between `python.ApplicationErrorType` and `py3.ApplicationErrorType`. This applices to the type enums for `ProtocolError` and `TransportError` as well.

Because we're using `cpdef enum` in `python`, it's not formatting as expected. This to me indicates we should migrate `lib.python` away from `cpdef enum` to have the formatting behavior expected for `enum.Enum` classes.

Reviewed By: yoney

Differential Revision: D55661465

fbshipit-source-id: 66df764bd542814441be175cc2701dbfd7231d79