Split extraction of base from IterInit* to IterBase
Summary:
Move extraction of iterator base to a separate bytecode. This way, IterInit*
will see either an array, even if the input was a legacy collection, or it will
see an object guaranteed to implement the Iterator interface.

This improves performance of foreach over legacy collections, as iterator
specialization optimizations will now see the underlying vec or dict.

After this diff, hhbbc will perform LIter optimization only when we were able
to elide the IterBase opcode. However, we can't update the LIterInit* logic
yet, as there is a similar LIter optimization in hackc, which is unable to
elide IterBase due to the lack of type information.

This diff is a pre-requisite for moving iterator base to a local, guarding on
it in both IterInit and IterNext, then specializing iteration per type.

Reviewed By: ricklavoie

Differential Revision: D55543001

fbshipit-source-id: ab6e8606396bfe5239302d10b4cc1076c4305f6a