Optimize iteration of Map/ImmMap/Set/ImmSet instead of ConstMap/ConstSet
Summary:
There seems to be a type monotonicity bug in one of the JIT optimizations
related to interfaces, where LdColDict is losing information about the
interface type.

When we emitted LdColDict, the base obviously had the interface type, but not
when LdColDict was getting lowered.

Revert this logic to check for the specific final classes while investigating
this further.

Differential Revision: D55764189

fbshipit-source-id: 365f5281afc9fd36de0347c7ef86beb751f1799e