Fix bug in reuse-immediate Vasm pass
Summary:
This pass could replace a ldimmq instruction, which doesn't set flags, with an
addqi instruction, which sets flags, and it wasn't checking if the there is a
live status-flags register at that point.  This diff fixes the problem by using an
lea instruction instead, which doesn't set the status flags.

Differential Revision: D55780056

fbshipit-source-id: 71029fcf51bb52958314e4c083535fe0b5063900