Retain errors through disjunctions
Summary:
Disjunctions are used in two ways from our constraint handling code.

First, we use the constructor to represent disjunctions we want to test

Second, we encode the `false` proposition as the empty disjunction, paired with an optional typing error, when we know the current proposition is false.

The second use means that whenever we create a disjunction, we provide the typing error of the current proposition that we are simplifying into a disjunction of propositions i.e. we retain the error related to the 'parent' proposition and generate this error is no disjunct is valid.

This diff modifies handling of disjunctions in two ways:
1) We explicitly handle the `false` proposition
2) We accumulate errors across non-empty disjunctions and return the disjunction of the errors

The diff also modifies the infix `|||` operator, used to construct non-empty disjunctions, so they no longer requires an optional typing error.

This improves error messages, particularly for cases where the one of the disjunct errors is a type parameter constraint violation and the inner error provides more context.

Reviewed By: andrewjkennedy

Differential Revision: D55641632

fbshipit-source-id: 93593f22e6820435e349518a25d5cbde09d37fd5