debugger_dump: don't print null
Summary:
## What
`debugger_dump` is used by Hack notebooks to print the last line of the notebook.

`debugger_dump` should not print `null` when given the argument `null`.

 https://pxl.cl/4Flvw

## Why:
- by user request (see task)
- for consistency with how Python notebooks work.

## More context

Note that users *can* write `debugger_dump` themselves, so this diff changes user-facing behavior. An alternative would be to do this logic in the [redeclaration daemon](https://www.internalfb.com/intern/wiki/Hack_Developer_Guide/) instead. We chose to put the logic in `debugger_dump` to aid maintainability: easiest if the logic is in one place.

Reviewed By: arnabde03

Differential Revision: D55575098

fbshipit-source-id: 8ae60d772a709f99fec006485136e4456c577292