Match on both sub- and supertype; inline C-Var-R cases
Summary:
As part of the provenance-aware subtyping work, we will be modifying simplification functions to only perform one step of a simplification. To make the correctness of the required changes clear, this modifies the top-level `match` to examine both sub- and supertype rather than using nested matches and wildcards

The diff also makes use of the match on subtypes to inline the, currently nested, matches for C-Var-R cases

Later in the stack we will de-duplicate the inlined code by driving the pattern match from the match subtype, in some cases (e.g. C-Var-L)

Reviewed By: mheiber

Differential Revision: D55795859

fbshipit-source-id: 64e82d76b0f9c4d5b3e84634e5de62e0bea11f56