Make wildcard subtype match explicit and inline `default_subtype` case for C-Bot-R
Summary:
As part of the provenance-aware subtyping work, we will be modifying simplification functions to only perform one step of a simplification.

The end-state for the refactor will be a pattern match with explicit subtype and supertype scrutinees with all heuristics and rewrites lifted out. The remaining match cases will then correspond exactly to a specification of constraint solving. The name C-Bot-R here refers to the (C)constraint solving case where we have the (B)bottom type on the (R)right hand side.

The diff makes the following changes for cases where we have `nothing` (the bottom type) in supertype position (C-Bot-R):
i) Make wildcard subtype match explicit
ii) Inline the `default_subtype` cases

Later in the stack we will de-duplicate the inlined code by driving the pattern match from the match subtype, in some cases (e.g. C-Var-L)

Differential Revision: D55796856

fbshipit-source-id: becbf68f216165219d3fe9c916af7f37355937b5