Unnest C-Shape-R subtype matches
Summary:
As part of the provenance-aware subtyping work, we will be modifying simplification functions to only perform one step of a simplification.

Moves matches for C-Shape-R case to the top-level.

Note that this is not _quite_ a refactor since it changes the handling of supportdyn. Previously we used the helper `strip_supportdyn` but this diff uses the existing subtyping rule to acheive the same effect

Differential Revision: D55921374

fbshipit-source-id: e8c7cdf13588c2fb9d4857ad245abc4f2d3dc84c