Use LIter instructions only with ArrayLike traversables
Summary:
If we are performing a LIter optimization, emit a check for ArrayLike input and
emit the foreach loop twice.

After this diff, LIter instruction will only observe ArrayLikes and objects
implementing the Iterator interface at runtime. This allows simplification of
LIter behavior so that it doesn't need to fall back to using the internal Iter
base.

Reviewed By: paulbiss

Differential Revision: D55856303

fbshipit-source-id: e2d25572e04a58d669d1595f2f8f26cd02eb1a33