Assume that LIter always receives extracted base
Summary:
After the previous diff, we generate LIter only with ArrayLikes, or when hhbbc
optimizes Iter into LIter, which it can do only when IterBase was optimized
away, which means that the base is extracted.

So assume in the interpreter that that's the case and assert that the base is
extracted rather than handling non-extracted base.

At this point JIT for both Iter and LIter instructions still handles
non-extracted bases. That logic will be completely refactor few diffs from now
and depends on the cleanup that this diff makes possible.

Reviewed By: ricklavoie

Differential Revision: D55856420

fbshipit-source-id: 12e483f02e8a1541de15997e44056882a0e6e647