Pass down owned ast::Program to emitter
Summary:
Now that hackc emitter backend has been refactored to use owned and interned
values instead of arena-allocated slices and references, it is natural to pass an owned
AST down to the emitter.  Individual owned strings can be passed to the final owned
Unit, or interned, and the AST can be consumed as we translate, instead of holding
the whole AST in memory the whole time. Over time this should lead to lower peak
memory use and more objects (mostly strings) moved into the final Unit instead of
being copied.

Reviewed By: voorka

Differential Revision: D55755172

fbshipit-source-id: d176c2ce9b370420a9d637bbb3976c89e7b6e0fe