factor out reprs
Summary:
hhbc::Body and ir::Func serve the same purpose, and have many of the
same metadata fields, but represent code differently.

* hhbc::Body - flat Vec of stack machine Opcode instructions with embedded Pseudo
instructions for SrcLoc and try/catch region markers.
* ir::Func - flat Vecs for SSA Instructions & immediates, and a vec of blocks forming a CFG, loc-per-instruction, and try/catch region info in side tables.

This diff scoops the fields that are different into struct BcRepr and IrRepr,
to get ready to dedup everything else.

Reviewed By: vassilmladenov

Differential Revision: D55916538

fbshipit-source-id: 160799aa501a9d80a8a409049583bb52998d36f6