Non-monotone static lint/error for class pointers
Summary:
This begins the process of breaking up `C::class : classname<T> <: string` in the typechecker. We raise a non-monotone error if a C::class expression flows to a position where the runtime will trigger a string conversion. Such an error is purely syntactic, and does not flow through e.g. variable assignment.

The logic is fairly primitive now, but I plan to extend the bool as needed into a ternary bumper of NoError, Lint, Error as needed for various types of expressions e.g. enum fields, dict keys, etc.

This is a precursor to changing the typing of `C::class` to `class<C> </: classname<C>`.

Reviewed By: viratyosin

Differential Revision: D53069663

fbshipit-source-id: 1693a35cdbbd849c485427d6c3112a4ce4f11987