Fix infinite loop bug in build_subclass_lists_assign
Summary:
When calculating splits, if one of the splits has exactly 25000
classes within it, we'll go into an infinite loop. This is because
such a split will also have the parent as a dependency, and that will
bump the size of 25001. The rest of the algorithm assumes that a valid
split can't be that large, so all sorts of things goes wrong. Fix this
by accounting for the extra class by using < instead of <= in
willProcess.

While I'm at it, do a bit of cleanup.

Reviewed By: mdko, voorka

Differential Revision: D56038171

fbshipit-source-id: 42df0f2d79fdce60f1aa7bb24d16e07e3b807bdb