introduce string flag rocket_frame_parser
Summary:
Current flags to manage which frame parser to use is rather error prone, as it is possible to set both `rocket_strategy_parser` and `rocket_allocating_strategy_parser` to be both true and it is ambiguous which one to use.

We are aiming for replace all these boolean flags with one single string flag rocket_frame_parser, this will make rolling out strategy parser a bit easier in the future.

Reviewed By: iahs

Differential Revision: D56035444

fbshipit-source-id: decee58317990a18634f14400743f100c2dd793f