Async jitting in sandbox mode
Summary:
This diff implements asynchronous jitting in background threads for live translations in sandbox mode. A request thread enqueues a translation request based on a `RegionContext` and continues in the interpreter. In order to avoid generating duplicate translations, only one translation request per `SrcKey` is  enqueued at a time. We also check if the number of translations for a `SrcKey` has changed between the time the request was enqueued and the translation attempt started and reject the translation request if it did. In order to reduce overlapping translations we do not enqueue translation requests after each basic block when `dispatchBB` returns.
We also generate function prologues using the same background threads.

Reviewed By: paulbiss

Differential Revision: D48879852

fbshipit-source-id: f0ddbb8753f9fcfb5937046e463387ba9c6d0466