Fix hhbc unresolved case type element
Summary:
Case type unions can't represent a mix of resolved and unresolved types. In the case where we attempt to resolve an unknown type in a union hhbbc was asserting instead of dealing with it gracefully - when an unresolved (unknown) type is present just treat the entire union as unresolved.

This change makes the hhbbc behavior match the hhvm sandbox behavior.

Note that there are a couple issues:
- A parameter with an unknown type as a union member will be treated as entirely unknown.  As a result a function with a signature like `int | Unknown` will be uncallable. But it may be better to ignore the unknown types as parameters and instead allow the function to be called with `int` and instead fail when/if we attempt to instantiate the unknown type.
- The error you get is not great - it refers to the original case type being missing as opposed naming exactly which subtypes are missing - which can be confusing. This isn't a huge deal because normally `hh` will have caught it first.

Reviewed By: ricklavoie

Differential Revision: D56143554

fbshipit-source-id: fe0bc5704f66bbcaad0010111da752f3ed22fd21