Require module imports to precede module exports
Summary:
The full-fidelity concrete syntax tree is intended to contain enough information to reproduce the entire input source file (even without position information), but the parsing for module imports and exports breaks that invariant. We are not sensitive to the ordering, and we represent a module declaration CST node as a struct with an exports member and an imports member (and no information about which came first).

This is a problem for hackfmt: the formatter is reliant upon this invariant for the behavior of range-formatting, which uses byte offsets (counted up as we iterate over EditableSyntax tokens, which contain no position information) to determine whether or not some token is in the range to be formatted. If we aren't sure whether exports or imports come first, we can't correctly count up these offsets, and range formatting breaks inside module declarations when imports precede exports.

If we want to allow exports and imports to be in any order, we should aggregate them in a SyntaxList. See D56159950 for an example of how we might do that parsing. I don't have time to finish that diff up, so I've submitted this simpler diff banning imports which follow exports.

Reviewed By: madgen

Differential Revision: D56161123

fbshipit-source-id: 6d354e4b6635b574e0ed5b07f9017c6689d42f8b