Refactor thrift python Field info data to be an actual type and add a field for the base non-annotated field name
Summary:
I was originally just going to add a value to the thrift python tuple to retain the base non annotated thrift name for a field, but it was maddeningly annoying to go and track down every single location where a tuple access happened, so I also migrated the tuple to be an actual object too.

This change needs to be made because JSON serialization uses the fields, so if you annotate a field with `python.Name` it breaks the serialization.

Reviewed By: yoney, ahilger

Differential Revision: D56091086

fbshipit-source-id: 26388c290875bd25695482d396bbcd8b534f9f6d