Use SignalSafeElfCache when printing C++ stack on crash
Summary:
Sometimes we crash during attempt to symbolize frames in C++ stack trace for
a crash.

Debugging seems to imply that some cached items stored in ElfCache went away
during its use.

Let's not use the shared ElfCache, but create a new one for this purpose. Since
there is an implementation that is safe to use in signal handlers, use it to
reduce potential interference with jemalloc that could have corrupted state.

Differential Revision: D56264145

fbshipit-source-id: c87f039b603d8666ac4c4a076aca50d8292408f8