Don't crash watchman when detected Eden root is not connected
Summary:
When watchman detects a root as an Eden root, but later dicovers that the Eden has stopped the mount or the mount is no longer connected - watchman SIGABRTs.

This happens on Linux, when Eden had a partially successful graceful takeover, but failed to start the FUSE channel (typically due to some authentication issues).

Lets make watchman more resliant to this situation by returning an TerminalWatcherError rather than passing through a std::system_exception. This will still fail the watchman command, but not crash the daemon - often resulting in a crash loop.

Reviewed By: kmancini

Differential Revision: D56043760

fbshipit-source-id: 728f862c0dd326fc75f6d38e80ad06fd117ccaa5