Fix `get_error_if_unsat` helper
Summary:
The `get_error_if_unsat` helper function is used to find the error attached to unsatisfiable propositions. Currently, the function has the signature:

```
val get_error_if_unsat: subtype_prop -> Typing_error.t option
```

This is used by the `with_error` and `&&&` helpers to combine errors for two unsatisfiable propositions into a single proposition contaning both errors.

The problem here is that we can obtain `None` from `get_error_if_unsat` even when we provide an unsatisfiable proposition as an argument since the contained `Typing_error.t option` will be also be `None` if no callback was supplied to the subtyping call.

In practice this will not cause errors to be dropped since we always retain one of the propositions.

To clarify the intent of the function, this diff modifies it to return a `Typing_error.t option` where `None` really does mean the prop was not unsatisfiable and `Some None` means it was but we didn't provide a callback.

Reviewed By: andrewjkennedy

Differential Revision: D56238051

fbshipit-source-id: 89f36b9514a4ebcb984e12c21aeceef3bde7a840