Remove `using namespace` from headers in hphp/facebook/extensions/carbon/...
Summary:
Putting `using namespace` in a header is a [bad anti-pattern](https://stackoverflow.com/questions/1452721/whats-the-problem-with-using-namespace-std) and we have multiple linters and compiler errors to prevent it. But to get global coverage of `-Wheader-hygiene`, we need to draw down existing instances.

This diff removes `using namespace` from one more header files, qualifying entities in the headers. Implementation files may gain `using namespace` which is _not_ a functional change: it merely makes explicit what was previously an implicit reliance on `using namespace`. It does, however, isolate the `using namespace` to just that implementation file, which is a significant improvement over the status quo.

Reviewed By: WizKid

Differential Revision: D56274521

fbshipit-source-id: a561732f0306f3d80a24b7fe6aaa1032cac9eeb6