Use boost validator for mode
Summary:
I was trying to figure out when to mark classes as persistent in SandboxAuthoritative mode, and
found it difficult to figure out what all the possible modes were and to trace through this function
to find where they are set.

I think this is better but am open to suggestions.

Add a custom validator for the ExecutionMode type.

Reviewed By: paulbiss

Differential Revision: D56287952

fbshipit-source-id: fadad4edbe45563ed7dd6e9346be39dd1209aabc